Add additional explanation for the code sample
Hi! very nice job, really enjoy following through...

So, the page "Converting Strings" has this code as example:

`fn main() -> Result<(), Box<dyn std::error::Error>> {
    let a = 42;
    let a_string = a.to_string();
    let b = a_string.parse::<i32>()?;
    println!("{} {}", a, b);
    Ok(())
}
`

and, it is the first time that. new concepts is introduced in the code sample but not 
explained on the left side... new concepts of Box, Trait (and syntax 'dyn Trait'). 
I don't think it was mentioned before.